Refactor chat bridging into a PlayersBridge
