+++ src/main/java/io/github/satoshinm/WebSandboxMC/bukkit/ChatListener.java
@@ -1,5 +1,6 @@
 package io.github.satoshinm.WebSandboxMC.bukkit;
 
+import io.github.satoshinm.WebSandboxMC.bridge.PlayersBridge;
 import io.github.satoshinm.WebSandboxMC.ws.WebSocketServerThread;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -7,18 +8,17 @@ import org.bukkit.event.player.AsyncPlayerChatEvent;
 
 public class ChatListener implements Listener {
 
-    private final WebSocketServerThread webSocketServerThread;
+    private final PlayersBridge playersBridge;
 
-    public ChatListener(WebSocketServerThread webSocketServerThread) {
-        this.webSocketServerThread = webSocketServerThread;
+    public ChatListener(PlayersBridge playersBridge) {
+        this.playersBridge = playersBridge;
     }
 
     @EventHandler(ignoreCancelled = true)
     public void onChat(AsyncPlayerChatEvent event) {
         String formattedMessage = event.getFormat().format(event.getMessage());
-        formattedMessage = "<" + event.getPlayer().getDisplayName() + "> " + formattedMessage; // TODO: why isn't getFormat() take care of this?
+        formattedMessage = "<" + event.getPlayer().getDisplayName() + "> " + formattedMessage; // TODO: why doesn't getFormat() take care of this?
 
-        // TODO: refactor logic out of thread into a ChatBridge
-        webSocketServerThread.notifyChat(formattedMessage);
+        playersBridge.notifyChat(formattedMessage);
     }
 }
